refactor(sim): relocate RiderBuilder from sim.rs to sim/rider.rs
RiderBuilder is the configuration surface for the rider lifecycle
(weight, group, route, patience, preferences, access control) and lives
naturally next to spawn_rider / build_rider, which already moved to
sim/rider.rs as part of the earlier sim.rs split. Keeping it in sim.rs
left the file holding 188 lines of rider-only code that referenced
super::Simulation through indirection.

Public path is preserved via `pub use rider::RiderBuilder;` in sim.rs
so external callers and the prelude re-export still see
elevator_core::sim::RiderBuilder. Internals downgrade from sim.rs's
private fields to pub(super) on the builder fields, since the spawn()
implementation lives in the rider submodule.

Audit §16.
